♻️ refactor: formatRelativeTime.ts 공통으로 분리
