--- src/app/(routers)/(board)/community/_utils/formatDate.ts
@@ -1,11 +1,0 @@
-export function formatDate(isoString: string): string {
-  const date = new Date(isoString);
-
-  if (isNaN(date.getTime())) return '';
-
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, '0');
-  const day = String(date.getDate()).padStart(2, '0');
-
-  return `${year}.${month}.${day}`;
-}
